+++ boards/frdmkv11z/demo_apps/mc_pmsm/mcdrv_adc_adc16.c
@@ -1,35 +1,9 @@
 /*
- * The Clear BSD License
  * Copyright (c) 2016, Freescale Semiconductor, Inc.
  * Copyright 2016-2018 NXP
  * All rights reserved.
  *
- * Redistribution and use in source and binary forms, with or without modification,
- * are permitted (subject to the limitations in the disclaimer below) provided
- * that the following conditions are met:
- *
- * o Redistributions of source code must retain the above copyright notice, this list
- *   of conditions and the following disclaimer.
- *
- * o Redistributions in binary form must reproduce the above copyright notice, this
- *   list of conditions and the following disclaimer in the documentation and/or
- *   other materials provided with the distribution.
- *
- * o Neither the name of the copyright holder nor the names of its
- *   contributors may be used to endorse or promote products derived from this
- *   software without specific prior written permission.
- *
- * NO EXPRESS OR IMPLIED LICENSES TO ANY PARTY'S PATENT RIGHTS ARE GRANTED BY THIS LICENSE.
- * THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS "AS IS" AND
- * ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE IMPLIED
- * WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE ARE
- * DISCLAIMED. IN NO EVENT SHALL THE COPYRIGHT HOLDER OR CONTRIBUTORS BE LIABLE FOR
- * ANY DIRECT, INDIRECT, INCIDENTAL, SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES
- * (INCLUDING, BUT NOT LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES;
- * LOSS OF USE, DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND ON
- * ANY THEORY OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT
- * (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF THIS
- * SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
+ * SPDX-License-Identifier: BSD-3-Clause
  */
 
 #include "mcdrv_adc_adc16.h"
@@ -115,20 +89,20 @@ bool_t MCDRV_Curr3Ph2ShChanAssignInit(mcdrv_adc16_t *this)
 
     /* update ADC_SC1n registers with initial values */
     this->sCurrSec16.pui32AdcBasePhaB->SC1[0] =
-        ((this->sCurrSec16.pui32AdcBasePhaB->SC1[0] & ~ADC_SC1_ADCH_MASK) | 
+        ((this->sCurrSec16.pui32AdcBasePhaB->SC1[0] & ~(uint16_t)ADC_SC1_ADCH_MASK) | 
           this->sCurrSec16.ui16ChanNumPhaB);
     this->sCurrSec16.pui32AdcBasePhaC->SC1[0] =
-        ((this->sCurrSec16.pui32AdcBasePhaC->SC1[0] & ~ADC_SC1_ADCH_MASK) | 
+        ((this->sCurrSec16.pui32AdcBasePhaC->SC1[0] & ~(uint16_t)ADC_SC1_ADCH_MASK) | 
            this->sCurrSec16.ui16ChanNumPhaC);
 
     /* Vdcb sampling - select SC1[1] (sample number 1) */
     this->pui32UdcbAdcBase->SC1[1] = 
-        ((this->pui32UdcbAdcBase->SC1[1] & ~ADC_SC1_ADCH_MASK) | 
+        ((this->pui32UdcbAdcBase->SC1[1] & ~(uint16_t)ADC_SC1_ADCH_MASK) | 
           this->ui16ChanNumVDcb);
     
     /* Aux sampling - select SC1[1] (sample number 1) */
     this->pui32AuxAdcBase->SC1[1] = 
-        ((this->pui32AuxAdcBase->SC1[1] & ~ADC_SC1_ADCH_MASK) | 
+        ((this->pui32AuxAdcBase->SC1[1] & ~(uint16_t)ADC_SC1_ADCH_MASK) | 
           this->ui16ChanNumAux);
 
     return (s_statusPass);
@@ -152,26 +126,26 @@ bool_t MCDRV_Curr3Ph2ShChanAssign(mcdrv_adc16_t *this)
         case 3: /* currents Ia, Ic will be measured */
             /* update ADC_SC1n registers based on actual SVM sector */
             this->sCurrSec23.pui32AdcBasePhaA->SC1[0] =
-                ((this->sCurrSec23.pui32AdcBasePhaA->SC1[0] & ~ADC_SC1_ADCH_MASK) | this->sCurrSec23.ui16ChanNumPhaA);
+                ((this->sCurrSec23.pui32AdcBasePhaA->SC1[0] & ~(uint16_t)ADC_SC1_ADCH_MASK) | this->sCurrSec23.ui16ChanNumPhaA);
             this->sCurrSec23.pui32AdcBasePhaC->SC1[0] =
-                ((this->sCurrSec23.pui32AdcBasePhaC->SC1[0] & ~ADC_SC1_ADCH_MASK) | this->sCurrSec23.ui16ChanNumPhaC);
+                ((this->sCurrSec23.pui32AdcBasePhaC->SC1[0] & ~(uint16_t)ADC_SC1_ADCH_MASK) | this->sCurrSec23.ui16ChanNumPhaC);
             break;
         case 4:
         case 5: /*  currents Ia, Ib will be measured */
             /* update ADC_SC1n registers based on actual SVM sector */
             this->sCurrSec45.pui32AdcBasePhaA->SC1[0] =
-                ((this->sCurrSec45.pui32AdcBasePhaA->SC1[0] & ~ADC_SC1_ADCH_MASK) | this->sCurrSec45.ui16ChanNumPhaA);
+                ((this->sCurrSec45.pui32AdcBasePhaA->SC1[0] & ~(uint16_t)ADC_SC1_ADCH_MASK) | this->sCurrSec45.ui16ChanNumPhaA);
             this->sCurrSec45.pui32AdcBasePhaB->SC1[0] =
-                ((this->sCurrSec45.pui32AdcBasePhaB->SC1[0] & ~ADC_SC1_ADCH_MASK) | this->sCurrSec45.ui16ChanNumPhaB);
+                ((this->sCurrSec45.pui32AdcBasePhaB->SC1[0] & ~(uint16_t)ADC_SC1_ADCH_MASK) | this->sCurrSec45.ui16ChanNumPhaB);
             break;
         case 1:
         case 6: /* currents Ib, Ic will be measured */
         default:
             /* update ADC_SC1n registers based on actual SVM sector */
             this->sCurrSec16.pui32AdcBasePhaB->SC1[0] =
-                ((this->sCurrSec16.pui32AdcBasePhaB->SC1[0] & ~ADC_SC1_ADCH_MASK) | this->sCurrSec16.ui16ChanNumPhaB);
+                ((this->sCurrSec16.pui32AdcBasePhaB->SC1[0] & ~(uint16_t)ADC_SC1_ADCH_MASK) | this->sCurrSec16.ui16ChanNumPhaB);
             this->sCurrSec16.pui32AdcBasePhaC->SC1[0] =
-                ((this->sCurrSec16.pui32AdcBasePhaC->SC1[0] & ~ADC_SC1_ADCH_MASK) | this->sCurrSec16.ui16ChanNumPhaC);
+                ((this->sCurrSec16.pui32AdcBasePhaC->SC1[0] & ~(uint16_t)ADC_SC1_ADCH_MASK) | this->sCurrSec16.ui16ChanNumPhaC);
             break;
     }
     return (s_statusPass);
